refactor: Tidy up date helpers
This change includes removing duplicate definitions for date formats
as well as updating file names to be a little more consistent within
the comment folder.
