+++ app/moves/middleware/set-body.single-requests.test.js
@@ -1,4 +1,4 @@
-const dateHelpers = require('../../../common/helpers/date-utils')
+const dateHelpers = require('../../../common/helpers/date')
 
 const middleware = require('./set-body.single-requests')
