+++ common/helpers/date.js
@@ -12,7 +12,6 @@ const {
 const { DATE_FORMATS } = require('../../config/index')
 
 module.exports = {
-  dateFormat: DATE_FORMATS.URL_PARAM,
   getCurrentWeekAsRange: () => {
     const today = new Date()
     const startDate = format(
@@ -37,21 +36,11 @@ module.exports = {
   },
   getRelativeDate: (date, interval) => {
     const method = interval >= 0 ? addDays : subDays
+
     return format(
       method(parseISO(date), Math.abs(interval)),
       DATE_FORMATS.URL_PARAM
     )
-  },
-  getDateFromParams: req => {
-    const date = req.params.date
-    const parsedDate = parseISO(date)
-    const validDate = isValidDate(parsedDate)
-
-    if (!validDate) {
-      return null
-    }
-
-    return format(parsedDate, DATE_FORMATS.URL_PARAM)
   },
   getDateRange: (date, timePeriod) => {
     const parsedDate = isDate(date) ? date : parseISO(date)
