+++ common/middleware/collection/index.js
@@ -2,6 +2,7 @@ const redirectDefaultQuery = require('./redirect-default-query')
 const redirectView = require('./redirect-view')
 const setActions = require('./set-actions')
 const setContext = require('./set-context')
+const setDateRange = require('./set-date-range')
 const setPagination = require('./set-pagination')
 
 module.exports = {
@@ -9,5 +10,6 @@ module.exports = {
   redirectView,
   setActions,
   setContext,
+  setDateRange,
   setPagination,
 }
