--- common/middleware/set-date-period.js
@@ -1,6 +1,0 @@
-function setDatePeriod(req, res, next, period) {
-  res.locals.period = period
-  next()
-}
-
-module.exports = setDatePeriod
